WebApi/Application/CustomerOperations/Commands/CreateUser/CreateUserCommand.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using AutoMapper;
using WebApi.DbOperations;
using WebApi.Entities;

namespace WebApi.Application.CustomerOperations.Commands.CreateUser
{
    public class CreateUserCommand
    {
        public CreateUserModel Model { get; set; }
        private readonly IMovieStoreDbContext _dbContext;
        private readonly IMapper _mapper;
        public CreateUserCommand(IMovieStoreDbContext dbContext, IMapper mapper)
        {
            _dbContext = dbContext;
            _mapper = mapper;
        }

        public void Handle()
        {
            var user =_dbContext.Customers.SingleOrDefault(x=> x.Email == Model.Email);
            if(user is not null)
            {
                throw new InvalidOperationException("Kullanici zaten mevcut.");
            }
            user = _mapper.Map<Customer>(Model);
            _dbContext.Customers.Add(user);
            _dbContext.SaveChanges();
        }
        public class CreateUserModel
        {
            public string Name { get; set; }
            public string Surname { get; set; }
            public string Email { get; set; }
            public string Password { get; set; }
        }
    }
}